refactored very big and scary urls.go into equally scary, but tiny, files
