+++ internal/web/urls.go
@@ -7,22 +7,12 @@ import (
 	"fmt"
 	"github.com/datasektionen/GOrdian/internal/config"
 	"github.com/datasektionen/GOrdian/internal/database"
-	"github.com/datasektionen/GOrdian/internal/excel"
 	"html/template"
-	"io"
 	"log/slog"
 	"math/rand"
 	"net/http"
 	"strconv"
 )
-
-type FrameLine struct {
-	FrameLineName     string
-	FrameLineIncome   int
-	FrameLineExpense  int
-	FrameLineInternal int
-	FrameLineResult   int
-}
 
 type Databases struct {
 	DBCF *sql.DB
@@ -70,25 +60,6 @@ func cors(h http.Handler) http.Handler {
 		w.Header().Add("Access-Control-Allow-Origin", "*")
 		h.ServeHTTP(w, r)
 	})
-}
-
-func add(x int, y int) int {
-	return x + y
-}
-
-func formatMoney(value int) string {
-	numStr := strconv.Itoa(value)
-	length := len(numStr)
-	var result string
-
-	for i := 0; i < length; i++ {
-		if i > 0 && (length-i)%3 == 0 {
-			result += " "
-		}
-		result += string(numStr[i])
-	}
-
-	return result
 }
 
 func route(db *sql.DB, handler func(w http.ResponseWriter, r *http.Request, db *sql.DB) error) http.Handler {
@@ -160,45 +131,23 @@ func sliceContains(list1 []string, list2 ...string) bool {
 	return false
 }
 
-func apiCostCentres(w http.ResponseWriter, r *http.Request, db *sql.DB) error {
-	costCentres, err := getCostCentres(db)
-	if err != nil {
-		return fmt.Errorf("failed get scan cost centres information from database: %v", err)
-	}
-	err = json.NewEncoder(w).Encode(costCentres)
-	if err != nil {
-		return fmt.Errorf("failed to encode cost centres to json: %v", err)
-	}
-	return nil
+func add(x int, y int) int {
+	return x + y
 }
 
-func apiSecondaryCostCentre(w http.ResponseWriter, r *http.Request, db *sql.DB) error {
-	idCC, err := strconv.Atoi(r.FormValue("id"))
-	if err != nil {
-		return fmt.Errorf("failed to convert secondary cost centre id to int: %v", err)
-	}
-	secondaryCostCentres, err := getSecondaryCostCentresByCostCentreID(db, idCC)
-	if err != nil {
-		return fmt.Errorf("failed get scan sendondary cost centres information from database: %v", err)
-	}
-	err = json.NewEncoder(w).Encode(secondaryCostCentres)
-	if err != nil {
-		return fmt.Errorf("failed to encode secondary cost centres to json: %v", err)
-	}
-	return nil
-}
+func formatMoney(value int) string {
+	numStr := strconv.Itoa(value)
+	length := len(numStr)
+	var result string
 
-func apiBudgetLine(w http.ResponseWriter, r *http.Request, db *sql.DB) error {
-	idSCC, err := strconv.Atoi(r.FormValue("id"))
-	budgetLines, err := getBudgetLinesBySecondaryCostCentreID(db, idSCC)
-	if err != nil {
-		return fmt.Errorf("failed get scan budget lines information from database: %v", err)
+	for i := 0; i < length; i++ {
+		if i > 0 && (length-i)%3 == 0 {
+			result += " "
+		}
+		result += string(numStr[i])
 	}
-	err = json.NewEncoder(w).Encode(budgetLines)
-	if err != nil {
-		return fmt.Errorf("failed to encode budget lines to json: %v", err)
-	}
-	return nil
+
+	return result
 }
 
 func adminPage(w http.ResponseWriter, r *http.Request, db *sql.DB, perms []string, loggedIn bool) error {
@@ -238,473 +187,6 @@ func logoutPage(w http.ResponseWriter, r *http.Request, db *sql.DB) error {
 	http.SetCookie(w, &sessionCookie)
 	http.Redirect(w, r, "/", http.StatusSeeOther)
 	return nil
-}
-
-func indexPage(w http.ResponseWriter, r *http.Request, db *sql.DB, perms []string, loggedIn bool) error {
-	costCentres, err := getCostCentres(db)
-	if err != nil {
-		return fmt.Errorf("failed get scan cost centre information from database: %v", err)
-	}
-	committeeCostCentres, projectCostCentres, otherCostCentres, err := splitCostCentresOnType(costCentres)
-	if err != nil {
-		return fmt.Errorf("failed to split cost centres on type: %v", err)
-	}
-
-	//Mörkläggning av mottagningens budget
-	darkeningResp, err := http.Get("https://darkmode.datasektionen.se/")
-	if err != nil {
-		slog.Error("Failed to get status from darkmode", "error", err)
-		return fmt.Errorf(": %v", err)
-	}
-	defer darkeningResp.Body.Close()
-
-	if darkeningResp.StatusCode != http.StatusOK {
-		slog.Error("Status error from darkmode", "error", darkeningResp.StatusCode)
-	}
-
-	darkeningBody, err := io.ReadAll(darkeningResp.Body)
-	if err != nil {
-		slog.Error("Failed to read body", "error", err)
-	}
-
-	darkeningValue, err := strconv.ParseBool(string(darkeningBody))
-	if err != nil {
-		slog.Error("Failed to parse bool", "error", err)
-	}
-
-	if darkeningValue {
-		for index, committeeCostCentre := range committeeCostCentres {
-			if committeeCostCentre.CostCentreName == "Mottagningen" {
-				committeeCostCentres = append(committeeCostCentres[:index], committeeCostCentres[index+1:]...)
-				break
-			}
-		}
-	}
-	//end of mörkläggning
-
-	if err := templates.ExecuteTemplate(w, "index.gohtml", map[string]any{
-		"motd":        motdGenerator(),
-		"committees":  committeeCostCentres,
-		"projects":    projectCostCentres,
-		"others":      otherCostCentres,
-		"permissions": perms,
-		"loggedIn":    loggedIn,
-	}); err != nil {
-		return fmt.Errorf("Could not render template: %w", err)
-	}
-	return nil
-}
-
-func framePage(w http.ResponseWriter, r *http.Request, db *sql.DB, perms []string, loggedIn bool) error {
-	budgetLines, err := getFrameLines(db)
-	if err != nil {
-		return fmt.Errorf("failed get scan budget lines information from database: %v", err)
-	}
-	committeeFrameLines, projectFrameLines, otherFrameLines, totalFrameLine, sumCommitteeFrameLine, sumProjectFrameLine, sumOtherFrameLine, err := generateFrameLines(budgetLines)
-	if err != nil {
-		return fmt.Errorf("failed to generate frame budget lines: %v", err)
-	}
-	if err := templates.ExecuteTemplate(w, "frame.gohtml", map[string]any{
-		"motd":                  motdGenerator(),
-		"committeeframelines":   committeeFrameLines,
-		"projectframelines":     projectFrameLines,
-		"otherframelines":       otherFrameLines,
-		"totalframeline":        totalFrameLine,
-		"sumcommitteeframeline": sumCommitteeFrameLine,
-		"sumprojectframeline":   sumProjectFrameLine,
-		"sumotherframeline":     sumOtherFrameLine,
-		"permissions":           perms,
-		"loggedIn":              loggedIn,
-	}); err != nil {
-		return fmt.Errorf("Could not render template: %w", err)
-	}
-	return nil
-}
-
-func reportPage(w http.ResponseWriter, r *http.Request, db *sql.DB, perms []string, loggedIn bool) error {
-
-	if err := templates.ExecuteTemplate(w, "report.gohtml", map[string]any{
-		"motd":        motdGenerator(),
-		"permissions": perms,
-		"loggedIn":    loggedIn,
-	}); err != nil {
-		return fmt.Errorf("Could not render template: %w", err)
-	}
-	return nil
-}
-
-func costCentrePage(w http.ResponseWriter, r *http.Request, db *sql.DB, perms []string, loggedIn bool) error {
-	costCentreIDString := r.PathValue("costCentreIDPath")
-	costCentreIDInt, err := strconv.Atoi(costCentreIDString)
-	if err != nil {
-		return fmt.Errorf("failed to convert cost centre id from string to int: %v", err)
-	}
-
-	budgetLines, err := getBudgetLinesByCostCentreID(db, costCentreIDInt)
-	if err != nil {
-		return fmt.Errorf("failed get scan budget line information from database: %v", err)
-	}
-
-	//omg
-	secondaryCostCentresWithBudgetLinesList := make([]secondaryCostCentresWithBudgetLines, 1)
-	currentSecondaryCostCentre := &secondaryCostCentresWithBudgetLinesList[0]
-	for _, budgetLine := range budgetLines {
-		if currentSecondaryCostCentre.SecondaryCostCentreName != budgetLine.SecondaryCostCentreName {
-			secondaryCostCentresWithBudgetLinesList = append(secondaryCostCentresWithBudgetLinesList, secondaryCostCentresWithBudgetLines{
-				SecondaryCostCentreName: budgetLine.SecondaryCostCentreName,
-				BudgetLines:             []excel.BudgetLine{},
-			})
-			currentSecondaryCostCentre = &secondaryCostCentresWithBudgetLinesList[len(secondaryCostCentresWithBudgetLinesList)-1]
-		}
-		currentSecondaryCostCentre.BudgetLines = append(currentSecondaryCostCentre.BudgetLines, budgetLine)
-	}
-	secondaryCostCentresWithBudgetLinesList = secondaryCostCentresWithBudgetLinesList[1:]
-
-	costCentre, err := getCostCentreByID(db, costCentreIDInt)
-	if err != nil {
-		return fmt.Errorf("failed get scan cost centre information from database: %v", err)
-	}
-
-	//calc the total incomes, expenses and results of all cost centres in the list
-	secondaryCostCentresWithBudgetLinesList, err = calculateSecondaryCostCentres(secondaryCostCentresWithBudgetLinesList)
-	if err != nil {
-		return fmt.Errorf("failed calculate secondary cost centre values: %v", err)
-	}
-
-	costCentreTotalIncome, costCentreTotalExpense, costCentreTotalResult, err := calculateCostCentre(secondaryCostCentresWithBudgetLinesList)
-	if err != nil {
-		return fmt.Errorf("failed calculate cost centre values: %v", err)
-	}
-
-	if err := templates.ExecuteTemplate(w, "costcentre.gohtml", map[string]any{
-		"motd": motdGenerator(),
-		"secondaryCostCentresWithBudgetLinesList": secondaryCostCentresWithBudgetLinesList,
-		"costCentre":             costCentre,
-		"costCentreTotalIncome":  costCentreTotalIncome,
-		"costCentreTotalExpense": costCentreTotalExpense,
-		"costCentreTotalResult":  costCentreTotalResult,
-		"permissions":            perms,
-		"loggedIn":               loggedIn,
-	}); err != nil {
-		return fmt.Errorf("could not render template: %w", err)
-	}
-	return nil
-}
-
-func calculateCostCentre(secondaryCostCentresWithBudgetLinesList []secondaryCostCentresWithBudgetLines) (int, int, int, error) {
-	var totalIncome int
-	var totalExpense int
-	for _, sCCWithBudgetLines := range secondaryCostCentresWithBudgetLinesList {
-		totalIncome = totalIncome + sCCWithBudgetLines.SecondaryCostCentreTotalIncome
-		totalExpense = totalExpense + sCCWithBudgetLines.SecondaryCostCentreTotalExpense
-	}
-	totalResult := totalIncome + totalExpense
-
-	return totalIncome, totalExpense, totalResult, nil
-}
-
-func calculateSecondaryCostCentres(secondaryCostCentresWithBudgetLinesList []secondaryCostCentresWithBudgetLines) ([]secondaryCostCentresWithBudgetLines, error) {
-	for index, sCCWithBudgetLines := range secondaryCostCentresWithBudgetLinesList {
-		var totalIncome int
-		var totalExpense int
-		for _, budgetLine := range sCCWithBudgetLines.BudgetLines {
-			totalIncome = totalIncome + budgetLine.BudgetLineIncome
-			totalExpense = totalExpense + budgetLine.BudgetLineExpense
-		}
-		secondaryCostCentresWithBudgetLinesList[index].SecondaryCostCentreTotalIncome = totalIncome
-		secondaryCostCentresWithBudgetLinesList[index].SecondaryCostCentreTotalExpense = totalExpense
-		secondaryCostCentresWithBudgetLinesList[index].SecondaryCostCentreTotalResult = totalIncome + totalExpense
-	}
-	return secondaryCostCentresWithBudgetLinesList, nil
-}
-
-type secondaryCostCentresWithBudgetLines struct {
-	SecondaryCostCentreName         string
-	SecondaryCostCentreTotalIncome  int
-	SecondaryCostCentreTotalExpense int
-	SecondaryCostCentreTotalResult  int
-	BudgetLines                     []excel.BudgetLine
-}
-
-func getBudgetLinesByCostCentreID(db *sql.DB, costCentreID int) ([]excel.BudgetLine, error) {
-	var budgetLinesGetStatementStatic = `
-		SELECT 
-    		budget_lines.id,
-    		budget_lines.name,
-    		income,
-			expense,
-			comment,
-			account,
-			secondary_cost_centres.id,
-			secondary_cost_centres.name
-		FROM budget_lines
-		JOIN secondary_cost_centres ON secondary_cost_centres.id = secondary_cost_centre_id
-		WHERE cost_centre_id = $1
-		ORDER BY secondary_cost_centre_id
-	`
-	result, err := db.Query(budgetLinesGetStatementStatic, costCentreID)
-	if err != nil {
-		return nil, fmt.Errorf("failed to get budget lines from database: %v", err)
-	}
-	var budgetLines []excel.BudgetLine
-	for result.Next() {
-		var budgetLine excel.BudgetLine
-
-		err := result.Scan(
-			&budgetLine.BudgetLineID,
-			&budgetLine.BudgetLineName,
-			&budgetLine.BudgetLineIncome,
-			&budgetLine.BudgetLineExpense,
-			&budgetLine.BudgetLineComment,
-			&budgetLine.BudgetLineAccount,
-			&budgetLine.SecondaryCostCentreID,
-			&budgetLine.SecondaryCostCentreName,
-		)
-		if err != nil {
-			return nil, fmt.Errorf("failed to scan budget line from query result: %v", err)
-		}
-		budgetLines = append(budgetLines, budgetLine)
-	}
-	return budgetLines, nil
-}
-
-func getSecondaryCostCentresByCostCentreID(db *sql.DB, costCentreID int) ([]excel.SecondaryCostCentre, error) {
-	var SecondaryCostCentresGetStatementStatic = `
-		SELECT 
-    		id,
-    		name,
-    		cost_centre_id
-		FROM secondary_cost_centres
-		WHERE cost_centre_id = $1
-		ORDER BY id
-	`
-	result, err := db.Query(SecondaryCostCentresGetStatementStatic, costCentreID)
-	if err != nil {
-		return nil, fmt.Errorf("failed to get secondary cost centres from database: %v", err)
-	}
-	var secondaryCostCentres []excel.SecondaryCostCentre
-	for result.Next() {
-		var secondaryCostCentre excel.SecondaryCostCentre
-
-		err := result.Scan(
-			&secondaryCostCentre.SecondaryCostCentreID,
-			&secondaryCostCentre.SecondaryCostCentreName,
-			&secondaryCostCentre.CostCentreID,
-		)
-		if err != nil {
-			return nil, fmt.Errorf("failed to scan secondary cost centre from query result: %v", err)
-		}
-		secondaryCostCentres = append(secondaryCostCentres, secondaryCostCentre)
-	}
-	return secondaryCostCentres, nil
-}
-
-func getBudgetLinesBySecondaryCostCentreID(db *sql.DB, secondaryCostCentreID int) ([]excel.BudgetLine, error) {
-	var budgetLinesGetStatementStatic = `
-		SELECT 
-    		id,
-    		name,
-    		income,
-			expense,
-			comment,
-			account,
-			secondary_cost_centre_id
-		FROM budget_lines
-		WHERE secondary_cost_centre_id = $1
-		ORDER BY id
-	`
-	result, err := db.Query(budgetLinesGetStatementStatic, secondaryCostCentreID)
-	if err != nil {
-		return nil, fmt.Errorf("failed to get budgetlines from database: %v", err)
-	}
-	var budgetLines []excel.BudgetLine
-	for result.Next() {
-		var budgetLine excel.BudgetLine
-
-		err := result.Scan(
-			&budgetLine.BudgetLineID,
-			&budgetLine.BudgetLineName,
-			&budgetLine.BudgetLineIncome,
-			&budgetLine.BudgetLineExpense,
-			&budgetLine.BudgetLineComment,
-			&budgetLine.BudgetLineAccount,
-			&budgetLine.SecondaryCostCentreID,
-		)
-		if err != nil {
-			return nil, fmt.Errorf("failed to scan budget line from query result: %v", err)
-		}
-		budgetLines = append(budgetLines, budgetLine)
-	}
-	return budgetLines, nil
-}
-
-func getFrameLines(db *sql.DB) ([]excel.BudgetLine, error) {
-	var frameLinesGetStatementStatic = `
-		SELECT 
-    		SUM(income),
-			SUM(expense),
-			secondary_cost_centres.name ILIKE '%Internt%',
-			cost_centres.id,
-			cost_centres.name,
-			cost_centres.type
-		FROM budget_lines
-		JOIN secondary_cost_centres ON secondary_cost_centres.id = secondary_cost_centre_id
-		JOIN cost_centres ON secondary_cost_centres.cost_centre_id = cost_centres.id
-		GROUP BY cost_centres.id, cost_centres.name, cost_centres.type, secondary_cost_centres.name ILIKE '%Internt%'
-		ORDER BY cost_centres.name, secondary_cost_centres.name ILIKE '%Internt%'
-	`
-	result, err := db.Query(frameLinesGetStatementStatic)
-	if err != nil {
-		return nil, fmt.Errorf("failed to get framelines from database: %v", err)
-	}
-	var frameLines []excel.BudgetLine
-	for result.Next() {
-		var frameLine excel.BudgetLine
-
-		err := result.Scan(
-			&frameLine.BudgetLineIncome,
-			&frameLine.BudgetLineExpense,
-			&frameLine.SecondaryCostCentreName,
-			&frameLine.CostCentreID,
-			&frameLine.CostCentreName,
-			&frameLine.CostCentreType,
-		)
-		if err != nil {
-			return nil, fmt.Errorf("failed to scan budget line from query result: %v", err)
-		}
-		frameLines = append(frameLines, frameLine)
-	}
-	return frameLines, nil
-}
-
-func getCostCentres(db *sql.DB) ([]excel.CostCentre, error) {
-	var costCentresGetStatementStatic = `SELECT id, name, type FROM cost_centres ORDER BY name`
-	result, err := db.Query(costCentresGetStatementStatic)
-	if err != nil {
-		return nil, fmt.Errorf("failed to get cost centres from database: %v", err)
-	}
-	var costCentres []excel.CostCentre
-	for result.Next() {
-		var costCentre excel.CostCentre
-
-		err := result.Scan(&costCentre.CostCentreID, &costCentre.CostCentreName, &costCentre.CostCentreType)
-		if err != nil {
-			return nil, fmt.Errorf("failed to scan cost centre from query result: %v", err)
-		}
-		costCentres = append(costCentres, costCentre)
-	}
-	return costCentres, nil
-}
-
-func getCostCentreByID(db *sql.DB, costCentreID int) (excel.CostCentre, error) {
-	var costCentreGetStatementStatic = `SELECT id, name, type FROM cost_centres WHERE id = $1`
-	result := db.QueryRow(costCentreGetStatementStatic, costCentreID)
-	var costCentre excel.CostCentre
-	err := result.Scan(&costCentre.CostCentreID, &costCentre.CostCentreName, &costCentre.CostCentreType)
-	if err != nil {
-		return excel.CostCentre{}, fmt.Errorf("failed to scan cost centre from query result: %v", err)
-	}
-	return costCentre, nil
-}
-
-func splitCostCentresOnType(costCentres []excel.CostCentre) ([]excel.CostCentre, []excel.CostCentre, []excel.CostCentre, error) {
-	var committeeCostCentres []excel.CostCentre
-	var projectCostCentres []excel.CostCentre
-	var otherCostCentres []excel.CostCentre
-	for _, costCentre := range costCentres {
-		switch costCentre.CostCentreType {
-		case "committee":
-			committeeCostCentres = append(committeeCostCentres, costCentre)
-		case "project":
-			projectCostCentres = append(projectCostCentres, costCentre)
-		case "other":
-			otherCostCentres = append(otherCostCentres, costCentre)
-		default:
-			return nil, nil, nil, fmt.Errorf("faulty cost centre type found when splitting")
-		}
-	}
-	return committeeCostCentres, projectCostCentres, otherCostCentres, nil
-}
-
-func generateFrameLines(frameLines []excel.BudgetLine) ([]FrameLine, []FrameLine, []FrameLine, FrameLine, FrameLine, FrameLine, FrameLine, error) {
-	var committeeFrameLines []FrameLine
-	var projectFrameLines []FrameLine
-	var otherFrameLines []FrameLine
-	var totalFrameLine FrameLine
-	var sumCommitteeFrameLine FrameLine
-	var sumProjectFrameLine FrameLine
-	var sumOtherFrameLine FrameLine
-
-	totalFrameLine.FrameLineName = "Totalt"
-	sumCommitteeFrameLine.FrameLineName = "Summa nämnder"
-	sumProjectFrameLine.FrameLineName = "Summa projekt"
-	sumOtherFrameLine.FrameLineName = "Summa övrigt"
-
-	var skippidi bool
-	for i, frameLine := range frameLines {
-		if skippidi == true {
-			skippidi = false
-			continue
-		}
-		frameLineIncome := frameLine.BudgetLineIncome
-		frameLineExpense := frameLine.BudgetLineExpense
-		frameLineName := frameLine.CostCentreName
-		frameLineInternal := 0
-		frameLineResult := 0
-
-		// each CC appears twice, once for internal costs, once for the rest
-		// both are handled i and i+1
-		// skippidi makes sure that the loop incements by two
-		if i+1 < len(frameLines) && frameLines[i+1].CostCentreName == frameLineName {
-			frameLineIncome += frameLines[i+1].BudgetLineIncome
-			frameLineExpense += frameLines[i+1].BudgetLineExpense
-			frameLineInternal = frameLines[i+1].BudgetLineIncome + frameLines[i+1].BudgetLineExpense
-			skippidi = true
-		}
-
-		frameLineResult = frameLineIncome + frameLineExpense
-
-		reconstructedFrameLine := FrameLine{frameLineName, frameLineIncome, frameLineExpense, frameLineInternal, frameLineResult}
-
-		totalFrameLine.FrameLineIncome += frameLineIncome
-		totalFrameLine.FrameLineExpense += frameLineExpense
-		totalFrameLine.FrameLineInternal += frameLineInternal
-		totalFrameLine.FrameLineResult += frameLineResult
-
-		switch frameLine.CostCentreType {
-		case "committee":
-			committeeFrameLines = append(committeeFrameLines, reconstructedFrameLine)
-		case "project":
-			projectFrameLines = append(projectFrameLines, reconstructedFrameLine)
-		case "other":
-			otherFrameLines = append(otherFrameLines, reconstructedFrameLine)
-		default:
-			return nil, nil, nil, FrameLine{}, FrameLine{}, FrameLine{}, FrameLine{}, fmt.Errorf("faulty cost centre type found when splitting")
-		}
-	}
-
-	for _, committeeFrameLine := range committeeFrameLines {
-		sumCommitteeFrameLine.FrameLineIncome += committeeFrameLine.FrameLineIncome
-		sumCommitteeFrameLine.FrameLineExpense += committeeFrameLine.FrameLineExpense
-		sumCommitteeFrameLine.FrameLineInternal += committeeFrameLine.FrameLineInternal
-		sumCommitteeFrameLine.FrameLineResult += committeeFrameLine.FrameLineResult
-	}
-
-	for _, ProjectFrameLine := range projectFrameLines {
-		sumProjectFrameLine.FrameLineIncome += ProjectFrameLine.FrameLineIncome
-		sumProjectFrameLine.FrameLineExpense += ProjectFrameLine.FrameLineExpense
-		sumProjectFrameLine.FrameLineInternal += ProjectFrameLine.FrameLineInternal
-		sumProjectFrameLine.FrameLineResult += ProjectFrameLine.FrameLineResult
-	}
-
-	for _, OtherFrameLine := range otherFrameLines {
-		sumOtherFrameLine.FrameLineIncome += OtherFrameLine.FrameLineIncome
-		sumOtherFrameLine.FrameLineExpense += OtherFrameLine.FrameLineExpense
-		sumOtherFrameLine.FrameLineInternal += OtherFrameLine.FrameLineInternal
-		sumOtherFrameLine.FrameLineResult += OtherFrameLine.FrameLineResult
-	}
-
-	return committeeFrameLines, projectFrameLines, otherFrameLines, totalFrameLine, sumCommitteeFrameLine, sumProjectFrameLine, sumOtherFrameLine, nil
 }
 
 func motdGenerator() string {
